Update AutonomousBrain.java
Hey guys, PLEASE PLEASE PLEASE measure the RR position for depositing with turn right outside of the warehouse. I realized I didnt save a copy of that since it was more frustrating for me to test, and I apologize. Also not feeling well. Thx sm and cya tmrw :)  Other than that the auton is streamlined and good to go.